fix: fak type and optional concrete columns and concrete beams
